scripts: simplify gp builder test by using directory diff

+++ tests/017.test.GlycoproteinBuilder.sh
@@ -16,21 +16,26 @@ echo -n "Testing 017.test.GlycoproteinBuilder.cpp ${prefix}..."
 rm -r "${directory}" >/dev/null 2>&1
 mkdir -p "${directory}"
 "${BIN_PATH}" "${input}" "${directory}" > "${directory}"/GlycoproteinBuilder.txt 2>&1
-fileList=("glycoprotein_initial.pdb" "glycoprotein.pdb" "0_glycoprotein.pdb" "1_glycoprotein.pdb" "glycoprotein.off" "glycoprotein_serialized.pdb" "GlycoproteinBuilder.txt")
-for file in "${fileList[@]}"; do
-    output="${directory}/${file}"
-    if [ ! -f "${output}" ]; then
-        echo -e "${prefix} Test FAILED!\n ${output} does not exist\n"
-        echo "Exit Code: 1"
-        return 1
-    fi
-    correct="tests/correct_outputs/${output}"
-    if ! cmp "${output}" "${correct}" >/dev/null 2>&1; then
-        echo -e "${prefix} Test FAILED!\n ${output} is different from ${correct}\n"
-        echo "Exit Code: 1"
-        return 1
-    fi
-done
+expected="tests/correct_outputs/${directory}"
+
+if [ ! -d "${expected}" ]; then
+    echo "Test FAILED"
+    echo "directory ${expected} does not exist"
+    echo "Exit Code: 1"
+    return 1
+fi
+# note: if diff starts being slow, consider comparing checksums of directory contents instead
+DIFF=$(diff -qr "${directory}" "${expected}")
+if [ "$DIFF" ]
+then
+    echo "Test FAILED"
+    echo "Difference between actual and expected output"
+    echo "Run the following command for details:"
+    echo "    diff -qr ${directory} ${expected}"
+    echo "Exit Code: 1"
+    return 1
+fi
+
 rm -r "${directory}" >/dev/null 2>&1
 printf "Test passed.\n"
 echo "Exit Code: 0"
